+++ src/SpreadsheetColumn.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Mivento\FilamentSpreadsheetEditor;
+
+use Mivento\FilamentSpreadsheetEditor\Builders\SpreadsheetColumn as Builder;
+
+class SpreadsheetColumn extends Builder
+{
+    //
+}