+++ src/SpreadsheetEditor.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Mivento\FilamentSpreadsheetEditor;
+
+use Mivento\FilamentSpreadsheetEditor\Builders\SpreadsheetEditor as Builder;
+
+class SpreadsheetEditor extends Builder
+{
+    //
+}